# app/public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manager API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/work/projects": {
            "get": {
                "tags": [
                    "Work Projects"
                ],
                "operationId": "a07df57323e00f38ca491a95255aa593",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "common"
                        ]
                    }
                ]
            }
        },
...
        "/work/tasks/{id}/plan": {
            "put": {
                "tags": [
                    "Work Task"
                ],
                "operationId": "176ff2fc08484e810ce164c802b5312a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "common"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Work Task"
                ],
                "operationId": "6f33714714863fad0ea2ce8385ea1b6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "common"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorModel": {
                "properties": {
                    "error": {
                        "nullable": true,
                        "properties": {
                            "code": {
                                "type": "integer"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "violations": {
                        "nullable": true,
                        "type": "array",
                        "items": {
                            "properties": {
                                "propertyPath": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "count": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "page": {
                        "type": "integer"
                    },
                    "pages": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "/authorize",
                        "scopes": {
                            "common": "Common"
                        }
                    }
                }
            }
        }
    }
}